requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.50.0
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
